More posting work

+++ electron-app/src/submission/submission.service.ts
@@ -4,6 +4,8 @@ import {
   BadRequestException,
   NotImplementedException,
   Logger,
+  forwardRef,
+  Inject,
 } from '@nestjs/common';
 import * as _ from 'lodash';
 import * as shortid from 'shortid';
@@ -37,6 +39,7 @@ export class SubmissionService {
     private fileSubmissionService: FileSubmissionService,
     private readonly validatorService: ValidatorService,
     private readonly eventEmitter: EventsGateway,
+    @Inject(forwardRef(() => PostService))
     private readonly postService: PostService,
   ) {}
 
@@ -86,6 +89,7 @@ export class SubmissionService {
       type: createDto.type,
       order: await this.repository.count(),
       created: Date.now(),
+      sources: [],
     };
 
     let completedSubmission = null;
@@ -134,6 +138,7 @@ export class SubmissionService {
   async deleteSubmission(id: string): Promise<number> {
     this.logger.log(id, 'Delete Submission');
     const submission = (await this.get(id)) as Submission;
+    this.postService.cancel(submission);
     switch (submission.type) {
       case SubmissionType.FILE:
         await this.fileSubmissionService.cleanupSubmission(submission as FileSubmission);
@@ -151,6 +156,11 @@ export class SubmissionService {
     this.logger.log(update.id, 'Update Submission');
     const { id, parts, postAt, removedParts } = update;
     const submissionToUpdate = (await this.get(id)) as Submission;
+
+    if (submissionToUpdate.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     submissionToUpdate.schedule.postAt = postAt;
     if (!postAt) {
       submissionToUpdate.schedule.isScheduled = false;
@@ -169,6 +179,10 @@ export class SubmissionService {
   async overwriteSubmissionParts(submissionOverwrite: SubmissionOverwrite) {
     const submission = (await this.get(submissionOverwrite.id, false)) as Submission;
 
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     const allParts = await this.partService.getPartsForSubmission(submission.id);
     const keepIds = submissionOverwrite.parts.map(p => p.accountId);
     const removeParts = allParts.filter(p => !keepIds.includes(p.accountId));
@@ -183,6 +197,10 @@ export class SubmissionService {
   async scheduleSubmission(id: string, isScheduled: boolean, postAt?: number): Promise<void> {
     this.logger.debug(`${id}: ${isScheduled}`, 'Schedule Submission');
     const submissionToSchedule = (await this.get(id)) as Submission;
+
+    if (submissionToSchedule.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
 
     if (postAt) {
       submissionToSchedule.schedule.postAt = postAt;
@@ -235,6 +253,9 @@ export class SubmissionService {
   async setPostAt(id: string, postAt: number | undefined): Promise<void> {
     this.logger.debug(`${id}: ${new Date(postAt).toLocaleString()}`, 'Update Submission Post At');
     const submission = (await this.get(id)) as Submission;
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
     submission.schedule.postAt = postAt;
     await this.repository.update(id, { schedule: submission.schedule });
     this.eventEmitter.emitOnComplete(
@@ -286,6 +307,11 @@ export class SubmissionService {
   async removeFileSubmissionThumbnail(id: string): Promise<SubmissionPackage<FileSubmission>> {
     this.logger.debug(id, 'Remove Submission Thumbnail');
     const submission: FileSubmission = (await this.get(id)) as FileSubmission;
+
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     const updated = await this.fileSubmissionService.removeThumbnail(submission);
     await this.repository.update(id, { thumbnail: null });
 
@@ -300,6 +326,11 @@ export class SubmissionService {
   ): Promise<SubmissionPackage<FileSubmission>> {
     this.logger.debug(location, 'Remove Submission Additional File');
     const submission: FileSubmission = (await this.get(id)) as FileSubmission;
+
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     const updated = await this.fileSubmissionService.removeAdditionalFile(submission, location);
     await this.repository.update(id, { additional: updated.additional });
 
@@ -315,6 +346,11 @@ export class SubmissionService {
   ): Promise<SubmissionPackage<FileSubmission>> {
     this.logger.debug(path, 'Change Submission Thumbnail');
     const submission: FileSubmission = (await this.get(id)) as FileSubmission;
+
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     const updated = await this.fileSubmissionService.changeThumbnailFile(submission, file, path);
     await this.repository.update(id, { thumbnail: updated.thumbnail });
 
@@ -329,6 +365,11 @@ export class SubmissionService {
     path: string,
   ): Promise<SubmissionPackage<FileSubmission>> {
     const submission: FileSubmission = (await this.get(id)) as FileSubmission;
+
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     const updated = await this.fileSubmissionService.changePrimaryFile(submission, file, path);
     await this.repository.update(id, { primary: updated.primary });
 
@@ -344,6 +385,11 @@ export class SubmissionService {
   ): Promise<SubmissionPackage<FileSubmission>> {
     this.logger.debug(path, 'Add Additional File');
     const submission: FileSubmission = (await this.get(id)) as FileSubmission;
+
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     const updated = await this.fileSubmissionService.addAdditionalFile(submission, file, path);
     await this.repository.update(id, { additional: updated.additional });
 
@@ -355,6 +401,11 @@ export class SubmissionService {
   async updateFileSubmissionAdditionalFile(id: string, record: FileRecord) {
     this.logger.debug(record, 'Updating Additional File');
     const submission: FileSubmission = (await this.get(id)) as FileSubmission;
+
+    if (submission.isPosting) {
+      throw new BadRequestException('Cannot update a submission that is posting');
+    }
+
     if (submission.additional) {
       const recordToUpdate: FileRecord = submission.additional.find(
         r => r.location === record.location,
